[Conv] Refractor to packages
