+++ fla/modules/conv/cp/__init__.py
@@ -1,0 +1,6 @@
+from .ops import CausalConv1dFunctionCP, causal_conv1d_cp
+
+__all__ = [
+    'CausalConv1dFunctionCP',
+    'causal_conv1d_cp',
+]
