+++ fla/modules/conv/cuda/__init__.py
@@ -1,0 +1,9 @@
+# Copyright (c) 2023-2025, Songlin Yang, Yu Zhang
+
+from .ops import FastCausalConv1dFn, causal_conv1d_cuda, fast_causal_conv1d_fn
+
+__all__ = [
+    'FastCausalConv1dFn',
+    'causal_conv1d_cuda',
+    'fast_causal_conv1d_fn',
+]
