+++ fla/modules/conv/triton/__init__.py
@@ -1,0 +1,19 @@
+# Copyright (c) 2023-2025, Songlin Yang, Yu Zhang
+
+from .ops import (
+    CausalConv1dFunction,
+    causal_conv1d_bwd,
+    causal_conv1d_fwd,
+    causal_conv1d_update,
+    causal_conv1d_update_states,
+    compute_dh0_triton,
+)
+
+__all__ = [
+    'CausalConv1dFunction',
+    'causal_conv1d_bwd',
+    'causal_conv1d_fwd',
+    'causal_conv1d_update',
+    'causal_conv1d_update_states',
+    'compute_dh0_triton',
+]
