+++ fla/modules/convolution.py
@@ -1,1845 +1,38 @@
 # Copyright (c) 2023-2025, Songlin Yang, Yu Zhang
 
-import math
-import warnings
 
-import torch
-import torch.distributed as dist
-import torch.nn as nn
-import torch.nn.functional as F
-import triton
-import triton.language as tl
-from einops import rearrange
-
-from fla.ops.cp import FLACPContext, conv_cp_send_recv_bwd, conv_cp_send_recv_fwd
-from fla.ops.utils import prepare_chunk_indices, prepare_sequence_ids
-from fla.utils import IS_AMD, autotune_cache_kwargs, get_multiprocessor_count, input_guard
-
-NUM_WARPS_AUTOTUNE = [2, 4, 8, 16] if IS_AMD else [4, 8, 16, 32]
-STATIC_WARPS = 32 if not IS_AMD else 16
-
-
-try:
-    from causal_conv1d import causal_conv1d_fn
-    from causal_conv1d import causal_conv1d_update as causal_conv1d_update_cuda
-    from causal_conv1d.cpp_functions import causal_conv1d_bwd_function
-except ImportError:
-    causal_conv1d_fn = None
-    causal_conv1d_update_cuda = None
-    causal_conv1d_bwd_function = None
-
-
-@triton.heuristics({
-    'HAS_WEIGHT': lambda args: args['weight'] is not None,
-    'HAS_BIAS': lambda args: args['bias'] is not None,
-    'HAS_RESIDUAL': lambda args: args['residual'] is not None,
-    'USE_INITIAL_STATE': lambda args: args['initial_state'] is not None,
-    'IS_VARLEN': lambda args: args['cu_seqlens'] is not None,
-})
-@triton.autotune(
-    configs=[
-        triton.Config({'BD': BD}, num_warps=num_warps)
-        for BD in [16, 32, 64, 128]
-        for num_warps in NUM_WARPS_AUTOTUNE
-    ],
-    key=['D', 'W', 'NB'],
-    **autotune_cache_kwargs,
+from fla.modules.conv import (
+    ImplicitLongConvolution,
+    LongConvolution,
+    PositionalEmbedding,
+    ShortConvolution,
+    causal_conv1d,
+    fft_conv,
 )
-@triton.jit
-def causal_conv1d_fwd_kernel(
-    x,
-    y,
-    weight,
-    bias,
-    residual,
-    cu_seqlens,
-    initial_state,
-    chunk_indices,
-    B,
-    T,
-    stride_x_n,
-    stride_x_t,
-    stride_x_d,
-    D: tl.constexpr,
-    W: tl.constexpr,
-    BT: tl.constexpr,
-    BW: tl.constexpr,
-    BD: tl.constexpr,
-    NB: tl.constexpr,
-    ACTIVATION: tl.constexpr,
-    HAS_WEIGHT: tl.constexpr,
-    HAS_BIAS: tl.constexpr,
-    HAS_RESIDUAL: tl.constexpr,
-    USE_INITIAL_STATE: tl.constexpr,
-    IS_VARLEN: tl.constexpr,
-):
-    i_d, i_t, i_b = tl.program_id(0), tl.program_id(1), tl.program_id(2)
-
-    if IS_VARLEN:
-        i_n, i_t = tl.load(chunk_indices + i_t * 2).to(tl.int32), tl.load(chunk_indices + i_t * 2 + 1).to(tl.int32)
-        bos, eos = tl.load(cu_seqlens + i_n).to(tl.int64), tl.load(cu_seqlens + i_n + 1).to(tl.int64)
-        T = eos - bos
-        p_x = x + bos * stride_x_t
-    else:
-        i_n = i_b
-        bos, eos = (i_b * T).to(tl.int64), (i_b * T + T).to(tl.int64)
-        p_x = x + i_b * stride_x_n
-
-    o_d = i_d * BD + tl.arange(0, BD)
-    o_w = tl.arange(0, BW) + W - BW
-    m_d = o_d < D
-    m_w = o_w >= 0
-
-    if HAS_WEIGHT:
-        # [BD, BW]
-        b_w = tl.load(weight + o_d[:, None] * W + o_w, mask=m_d[:, None] & m_w, other=0).to(tl.float32)
-
-    b_y = tl.zeros((BT, BD), dtype=tl.float32)
-    if not USE_INITIAL_STATE:
-        for i_w in tl.static_range(-W + 1, 1):
-            p_yi = tl.make_block_ptr(p_x, (T, D), (stride_x_t, stride_x_d), (i_t * BT + i_w, i_d * BD), (BT, BD), (1, 0))
-            # [BT, BD]
-            b_yi = tl.load(p_yi, boundary_check=(0, 1)).to(tl.float32)
-            if HAS_WEIGHT:
-                b_yi *= tl.sum(b_w * (o_w == (i_w + W - 1)), 1)
-            b_y += b_yi
-    elif i_t * BT >= W:
-        # to make Triton compiler happy, we need to copy codes
-        for i_w in tl.static_range(-W + 1, 1):
-            p_yi = tl.make_block_ptr(p_x, (T, D), (stride_x_t, stride_x_d), (i_t * BT + i_w, i_d * BD), (BT, BD), (1, 0))
-            # [BT, BD]
-            b_yi = tl.load(p_yi, boundary_check=(0, 1)).to(tl.float32)
-            if HAS_WEIGHT:
-                b_yi *= tl.sum(b_w * (o_w == (i_w + W - 1)), 1)
-            b_y += b_yi
-    else:
-        o_t = i_t * BT + tl.arange(0, BT)
-        for i_w in tl.static_range(-W + 1, 1):
-            o_x = o_t + i_w
-            m_x = ((o_x >= 0) & (o_x < T))[:, None] & m_d
-            m_c = ((o_x + W >= 0) & (o_x < 0))[:, None] & m_d
-
-            b_yi = tl.load(
-                p_x + o_x[:, None] * stride_x_t + o_d * stride_x_d,
-                mask=m_x,
-                other=0
-            ).to(tl.float32)
-
-            b_yi += tl.load(initial_state + i_n * D*W + o_d * W + (o_x + W)[:, None], mask=m_c, other=0).to(tl.float32)
-
-            if HAS_WEIGHT:
-                b_yi *= tl.sum(b_w * (o_w == (i_w + W - 1)), 1)
-            b_y += b_yi
-
-    if HAS_BIAS:
-        b_y += tl.load(bias + o_d, mask=m_d).to(tl.float32)
-
-    if ACTIVATION == 'swish' or ACTIVATION == 'silu':
-        b_y = b_y * tl.sigmoid(b_y)
-
-    if HAS_RESIDUAL:
-        p_residual = tl.make_block_ptr(residual + bos * D, (T, D), (D, 1), (i_t * BT, i_d * BD), (BT, BD), (1, 0))
-        b_residual = tl.load(p_residual, boundary_check=(0, 1))
-        b_y += b_residual
-
-    p_y = tl.make_block_ptr(y + bos * D, (T, D), (D, 1), (i_t * BT, i_d * BD), (BT, BD), (1, 0))
-    tl.store(p_y, tl.cast(b_y, dtype=p_y.dtype.element_ty, fp_downcast_rounding='rtne'), boundary_check=(0, 1))
-
-
-@triton.heuristics({
-    'HAS_WEIGHT': lambda args: args['dw'] is not None,
-    'HAS_BIAS': lambda args: args['db'] is not None,
-    'USE_INITIAL_STATE': lambda args: args['initial_state'] is not None,
-    'USE_FINAL_STATE': lambda args: args['dht'] is not None,
-    'IS_VARLEN': lambda args: args['cu_seqlens'] is not None,
-})
-@triton.autotune(
-    configs=[
-        triton.Config({'BD': BD}, num_warps=num_warps)
-        for BD in [16, 32, 64, 128]
-        for num_warps in [4, 8, 16, 32]
-    ],
-    key=['D', 'W', 'NB'],
-    **autotune_cache_kwargs,
+from fla.modules.conv.cp import CausalConv1dFunctionCP, causal_conv1d_cp
+from fla.modules.conv.cuda import FastCausalConv1dFn, fast_causal_conv1d_fn
+from fla.modules.conv.triton import (
+    CausalConv1dFunction,
+    causal_conv1d_bwd,
+    causal_conv1d_fwd,
+    causal_conv1d_update,
+    causal_conv1d_update_states,
 )
-@triton.jit
-def causal_conv1d_bwd_kernel(
-    x,
-    y,
-    weight,
-    initial_state,
-    dht,
-    dy,
-    dx,
-    dw,
-    db,
-    cu_seqlens,
-    chunk_indices,
-    B,
-    T,
-    stride_x_n,   # x batch stride
-    stride_x_t,   # x time stride
-    stride_x_d,   # x dim stride
-    stride_dx_n,  # dx batch stride
-    stride_dx_t,  # dx time stride
-    stride_dx_d,  # dx dim stride
-    D: tl.constexpr,
-    W: tl.constexpr,
-    BT: tl.constexpr,
-    BW: tl.constexpr,
-    BD: tl.constexpr,
-    NB: tl.constexpr,
-    ACTIVATION: tl.constexpr,
-    HAS_WEIGHT: tl.constexpr,
-    HAS_BIAS: tl.constexpr,
-    USE_INITIAL_STATE: tl.constexpr,
-    USE_FINAL_STATE: tl.constexpr,
-    IS_VARLEN: tl.constexpr,
-):
-    i_d, i_t, i_b = tl.program_id(0), tl.program_id(1), tl.program_id(2)
-    if IS_VARLEN:
-        i_tg = i_t
-        i_n, i_t = tl.load(chunk_indices + i_t * 2).to(tl.int32), tl.load(chunk_indices + i_t * 2 + 1).to(tl.int32)
-        bos, eos = tl.load(cu_seqlens + i_n).to(tl.int64), tl.load(cu_seqlens + i_n + 1).to(tl.int64)
-        T = eos - bos
-        p_x = x + bos * stride_x_t
-    else:
-        i_tg = i_b * tl.num_programs(1) + i_t
-        i_n = i_b
-        bos, eos = (i_b * T).to(tl.int64), (i_b * T + T).to(tl.int64)
-        p_x = x + i_b * stride_x_n
 
-    o_d = i_d * BD + tl.arange(0, BD)
-    o_w = tl.arange(0, BW) + W - BW
-    m_d = o_d < D
-    m_w = o_w >= 0
-
-    if HAS_WEIGHT:
-        p_x = tl.make_block_ptr(p_x, (T, D), (stride_x_t, stride_x_d), (i_t * BT, i_d * BD), (BT, BD), (1, 0))
-        b_x = tl.load(p_x, boundary_check=(0, 1))
-        # [BD, BW]
-        b_w = tl.load(weight + o_d[:, None] * W + o_w, mask=m_d[:, None] & m_w, other=0)
-
-    b_dx = tl.zeros((BT, BD), dtype=tl.float32)
-    if HAS_BIAS:
-        b_db = tl.zeros((BD,), dtype=tl.float32)
-
-    if not USE_FINAL_STATE and not USE_INITIAL_STATE:
-        for i_w in tl.static_range(0, W):
-            p_dy = tl.make_block_ptr(dy + bos * D, (T, D), (D, 1), (i_t * BT + i_w, i_d * BD), (BT, BD), (1, 0))
-            # [BT, BD]
-            b_dy = tl.load(p_dy, boundary_check=(0, 1)).to(tl.float32)
-            if ACTIVATION == 'swish' or ACTIVATION == 'silu':
-                p_y = tl.make_block_ptr(y + bos * D, (T, D), (D, 1), (i_t * BT + i_w, i_d * BD), (BT, BD), (1, 0))
-                b_y = tl.load(p_y, boundary_check=(0, 1)).to(tl.float32)
-                b_ys = tl.sigmoid(b_y)
-                b_dy = b_dy * b_ys * (1 + b_y * (1 - b_ys))
-            b_wdy = b_dy
-            if HAS_WEIGHT:
-                # [BT, BD]
-                b_wdy = b_wdy * tl.sum(b_w * (o_w == (W - i_w - 1)), 1)
-                # [BD]
-                b_dw = tl.sum(b_dy * b_x, 0)
-                tl.store(dw + i_tg * D*W + o_d * W + W - i_w - 1, b_dw.to(dw.dtype.element_ty), mask=m_d)
-            if HAS_BIAS and i_w == 0:
-                b_db += tl.sum(b_dy, 0)
-            b_dx += b_wdy
-    elif i_t * BT >= W:
-        # to make Triton compiler happy, we need to copy codes
-        for i_w in tl.static_range(0, W):
-            p_dy = tl.make_block_ptr(dy + bos * D, (T, D), (D, 1), (i_t * BT + i_w, i_d * BD), (BT, BD), (1, 0))
-            # [BT, BD]
-            b_dy = tl.load(p_dy, boundary_check=(0, 1)).to(tl.float32)
-            if ACTIVATION == 'swish' or ACTIVATION == 'silu':
-                p_y = tl.make_block_ptr(y + bos * D, (T, D), (D, 1), (i_t * BT + i_w, i_d * BD), (BT, BD), (1, 0))
-                b_y = tl.load(p_y, boundary_check=(0, 1)).to(tl.float32)
-                b_ys = tl.sigmoid(b_y)
-                b_dy = b_dy * b_ys * (1 + b_y * (1 - b_ys))
-            b_wdy = b_dy
-            if HAS_WEIGHT:
-                # [BT, BD]
-                b_wdy = b_wdy * tl.sum(b_w * (o_w == (W - i_w - 1)), 1)
-                # [BD]
-                b_dw = tl.sum(b_dy * b_x, 0)
-                tl.store(dw + i_tg * D*W + o_d * W + W - i_w - 1, b_dw.to(dw.dtype.element_ty), mask=m_d)
-            if HAS_BIAS and i_w == 0:
-                b_db += tl.sum(b_dy, 0)
-            b_dx += b_wdy
-    else:
-        # which may use initial state
-        o_t = i_t * BT + tl.arange(0, BT)
-        for i_w in tl.static_range(0, W):
-            p_dy = tl.make_block_ptr(dy + bos * D, (T, D), (D, 1), (i_t * BT + i_w, i_d * BD), (BT, BD), (1, 0))
-            b_dy_shift = tl.load(p_dy, boundary_check=(0, 1)).to(tl.float32)
-            if ACTIVATION == 'swish' or ACTIVATION == 'silu':
-                p_y = tl.make_block_ptr(y + bos * D, (T, D), (D, 1), (i_t * BT + i_w, i_d * BD), (BT, BD), (1, 0))
-                b_y_shift = tl.load(p_y, boundary_check=(0, 1)).to(tl.float32)
-                b_ys = tl.sigmoid(b_y_shift)
-                b_dy_shift = b_dy_shift * b_ys * (1 + b_y_shift * (1 - b_ys))
-            if HAS_WEIGHT:
-                # gradient comes from x：sum_t dy[t+i_w] * x[t]
-                b_dw = tl.sum(b_dy_shift * b_x, 0)
-                # index of cache：c = W - i_w + t
-                if USE_INITIAL_STATE:
-                    mask_head_rows = (o_t < i_w)
-                    # dy_head = dy[t]
-                    b_dy_head = tl.load(dy + bos * D + o_t[:, None] * D + o_d, mask=(mask_head_rows[:, None] & m_d[None, :]),
-                                        other=0.0).to(tl.float32)
-                    if ACTIVATION == 'swish' or ACTIVATION == 'silu':
-                        # use y[t] （not y[t+i_w]）
-                        b_y_head = tl.load(y + bos * D + o_t[:, None] * D + o_d,
-                                           mask=(mask_head_rows[:, None] & m_d[None, :]), other=0.0).to(tl.float32)
-                        b_ys_head = tl.sigmoid(b_y_head)
-                        b_dy_head = b_dy_head * b_ys_head * (1 + b_y_head * (1 - b_ys_head))
-                    o_c = W - i_w + o_t
-                    # index 0 is padding 0
-                    mask_c = (mask_head_rows & (o_c >= 1) & (o_c < W))
-                    b_xc = tl.load(initial_state + i_n * D * W + o_d[None, :] * W + o_c[:, None],
-                                   mask=(mask_c[:, None] & m_d[None, :]), other=0.0).to(tl.float32)
-                    # add the gradient comes from initial_state
-                    b_dw += tl.sum(b_dy_head * b_xc, 0)
-                tl.store(dw + i_tg * D * W + o_d * W + W - i_w - 1, b_dw.to(dw.dtype.element_ty), mask=m_d)
-
-            if HAS_BIAS and i_w == 0:
-                b_db += tl.sum(b_dy_shift, 0)
-            b_wdy = b_dy_shift if not HAS_WEIGHT else (b_dy_shift * tl.sum(b_w * (o_w == (W - i_w - 1)), 1))
-            b_dx += b_wdy
-
-    if HAS_BIAS:
-        b_db = tl.cast(b_db, dtype=db.dtype.element_ty, fp_downcast_rounding='rtne')
-        tl.store(db + i_tg * D + o_d, b_db, mask=m_d)
-
-    if USE_FINAL_STATE:
-        if i_t * BT + BT >= T-W:
-            start_tok = max(0, T - (W - 1))
-            offset = i_t * BT + tl.arange(0, BT)
-            tok_idx = offset - start_tok
-            mask = (offset >= start_tok) & (offset < T)
-            w_idx = 1 + tok_idx
-            dht_off = i_n * D * W + o_d[None, :] * W + w_idx[:, None]
-            b_dht = tl.load(dht + dht_off, mask=mask[:, None] & m_d[None, :], other=0.).to(tl.float32)
-            b_dx += b_dht
-
-    if IS_VARLEN:
-        p_dx = dx + bos * stride_dx_t
-    else:
-        p_dx = dx + i_b * stride_dx_n
-
-    p_dx = tl.make_block_ptr(p_dx, (T, D), (stride_dx_t, stride_dx_d), (i_t * BT, i_d * BD), (BT, BD), (1, 0))
-    tl.store(p_dx, tl.cast(b_dx, dtype=p_dx.dtype.element_ty, fp_downcast_rounding='rtne'), boundary_check=(0, 1))
-
-
-@triton.heuristics({
-    'USE_INITIAL_STATE': lambda args: args['cache'] is not None,
-    'HAS_WEIGHT': lambda args: args['weight'] is not None,
-    'HAS_BIAS': lambda args: args['bias'] is not None,
-    'HAS_RESIDUAL': lambda args: args['residual'] is not None,
-})
-@triton.jit
-def causal_conv1d_update_kernel(
-    x,
-    cache,
-    residual,
-    y,
-    weight,
-    bias,
-    stride_x_n,  # batch stride
-    stride_x_d,  # dim stride
-    stride_y_n,  # batch stride
-    stride_y_d,  # dim stride
-    D: tl.constexpr,
-    W: tl.constexpr,
-    BD: tl.constexpr,
-    BW: tl.constexpr,
-    ACTIVATION: tl.constexpr,
-    USE_INITIAL_STATE: tl.constexpr,
-    HAS_WEIGHT: tl.constexpr,
-    HAS_BIAS: tl.constexpr,
-    HAS_RESIDUAL: tl.constexpr,
-):
-    i_d, i_n = tl.program_id(0), tl.program_id(1)
-
-    o_d = i_d * BD + tl.arange(0, BD)
-    o_w = tl.arange(0, BW)
-    m_d = o_d < D
-    m_w = o_w < W
-
-    # [BD]
-    b_x = tl.load(x + i_n * stride_x_n + o_d * stride_x_d, mask=m_d, other=0).to(tl.float32)
-
-    b_cache = tl.zeros((BD, BW), dtype=tl.float32)
-
-    if USE_INITIAL_STATE:
-        # 2. Shift Cache (Read [1:])
-        p_cache_read = tl.make_block_ptr(
-            cache + i_n * D*W,
-            shape=(D, W),
-            strides=(W, 1),
-            offsets=(i_d * BD, 1),
-            block_shape=(BD, BW),
-            order=(1, 0)
-        )
-        b_cache = tl.load(p_cache_read, boundary_check=(0, 1)).to(tl.float32)
-
-        # 3. Fill x to the last position
-        m_update = o_w == (W - 1)
-        b_cache = tl.where(m_update[None, :], b_x[:, None], b_cache)
-
-    if HAS_WEIGHT:
-        b_w = tl.load(weight + o_d[:, None] * W + o_w, mask=m_d[:, None] & m_w, other=0)
-        b_y = tl.sum(b_cache * b_w, 1)
-    else:
-        b_y = tl.sum(b_cache, 1)
-
-    if HAS_BIAS:
-        b_y += tl.load(bias + o_d, mask=m_d)
-
-    if ACTIVATION == 'swish' or ACTIVATION == 'silu':
-        b_y = b_y * tl.sigmoid(b_y)
-
-    if HAS_RESIDUAL:
-        b_y += tl.load(residual + i_n * D + o_d, mask=m_d, other=0)
-
-    tl.store(y + i_n * stride_y_n + o_d * stride_y_d, tl.cast(b_y,
-             dtype=y.dtype.element_ty, fp_downcast_rounding='rtne'), mask=m_d)
-
-    if USE_INITIAL_STATE:
-        p_cache_write = tl.make_block_ptr(
-            cache + i_n * D*W,
-            shape=(D, W),
-            strides=(W, 1),
-            offsets=(i_d * BD, 0),
-            block_shape=(BD, BW),
-            order=(1, 0)
-        )
-        tl.store(p_cache_write, tl.cast(b_cache, dtype=cache.dtype.element_ty,
-                 fp_downcast_rounding='rtne'), boundary_check=(0, 1))
-
-
-@input_guard(no_guard_contiguous=["x"])
-def causal_conv1d_fwd(
-    x: torch.Tensor,
-    weight: torch.Tensor,
-    bias: torch.Tensor,
-    residual: torch.Tensor,
-    initial_state: torch.Tensor | None = None,
-    output_final_state: bool = False,
-    activation: str | None = None,
-    cu_seqlens: torch.LongTensor | None = None,
-    cu_seqlens_cpu: torch.LongTensor | None = None,
-    chunk_indices: torch.LongTensor | None = None,
-) -> torch.Tensor:
-    shape = x.shape
-    if x.shape[-1] != weight.shape[0]:
-        x = rearrange(x, 'b t ... -> b t (...)')
-    B, T, D = x.shape[0], x.shape[1], weight.shape[0]
-    W = weight.shape[1]
-    stride_x_n, stride_x_t, stride_x_d = x.stride()
-
-    BT = min(64, triton.next_power_of_2(triton.cdiv(max(16, B*T), get_multiprocessor_count(x.device.index))))
-    BW = triton.next_power_of_2(W)
-    if chunk_indices is None and (cu_seqlens is not None or cu_seqlens_cpu is not None):
-        chunk_indices = prepare_chunk_indices(cu_seqlens, BT, cu_seqlens_cpu=cu_seqlens_cpu)
-    NT = len(chunk_indices) if cu_seqlens is not None else triton.cdiv(T, BT)
-    NB = triton.cdiv(B*T, 1024)
-
-    y = torch.empty_like(x, memory_format=torch.contiguous_format)
-
-    def grid(meta): return (triton.cdiv(D, meta['BD']), NT, B)
-    causal_conv1d_fwd_kernel[grid](
-        x=x,
-        y=y,
-        weight=weight,
-        bias=bias,
-        residual=residual,
-        cu_seqlens=cu_seqlens,
-        initial_state=initial_state,
-        chunk_indices=chunk_indices,
-        B=B,
-        T=T,
-        D=D,
-        W=W,
-        BT=BT,
-        BW=BW,
-        NB=NB,
-        stride_x_n=stride_x_n,
-        stride_x_t=stride_x_t,
-        stride_x_d=stride_x_d,
-        ACTIVATION=activation,
-    )
-    final_state = None
-    if output_final_state:
-        final_state = causal_conv1d_update_states(
-            x=x,
-            state_len=W,
-            initial_state=initial_state,
-            cu_seqlens=cu_seqlens,
-        )
-    return y.view(shape), final_state
-
-
-@triton.heuristics({
-    'USE_ACTIVATION': lambda args: args['y'] is not None,
-    'IS_VARLEN': lambda args: args['cu_seqlens'] is not None,
-})
-@triton.jit
-def compute_dh0_kernel(
-    dy,
-    y,
-    weight,
-    dh0,
-    cu_seqlens,
-    stride_dy_n,
-    stride_dy_t,
-    T,
-    D: tl.constexpr,
-    W: tl.constexpr,
-    BD: tl.constexpr,
-    USE_ACTIVATION: tl.constexpr,
-    IS_VARLEN: tl.constexpr,
-):
-    """
-    Compute dh0 (gradient w.r.t. initial_state) in a separate kernel.
-    This avoids Triton compiler bugs on some architectures (e.g., GB200).
-
-    Grid: (cdiv(D, BD), N)
-    """
-    i_d, i_n = tl.program_id(0), tl.program_id(1)
-
-    # Get sequence boundaries
-    if IS_VARLEN:
-        bos = tl.load(cu_seqlens + i_n).to(tl.int32)
-        eos = tl.load(cu_seqlens + i_n + 1).to(tl.int32)
-        seq_len = eos - bos
-        # For varlen, dy is [1, total_T, D], offset by bos
-        dy_base = dy + bos * stride_dy_t
-    else:
-        seq_len = T
-        # For non-varlen, dy is [B, T, D], offset by i_n * stride_dy_n
-        dy_base = dy + i_n * stride_dy_n
-
-    o_d = i_d * BD + tl.arange(0, BD)
-    m_d = o_d < D
-
-    # For each i_w in [1, W), compute dh0[i_n, :, i_w]
-    for i_w in tl.static_range(1, W):
-        b_dh0 = tl.zeros([BD], dtype=tl.float32)
-
-        # Accumulate contributions from t = 0 to min(i_w, seq_len) - 1
-        for t in tl.static_range(0, W - 1):
-            if t < i_w:
-                w_idx = i_w - 1 - t
-
-                # Load dy[t, :] relative to dy_base
-                p_dy = dy_base + t * stride_dy_t + o_d
-                m_t = (t < seq_len) & m_d
-                b_dy = tl.load(p_dy, mask=m_t, other=0).to(tl.float32)
-
-                if USE_ACTIVATION:
-                    if IS_VARLEN:
-                        p_y = y + bos * stride_dy_t + t * stride_dy_t + o_d
-                    else:
-                        p_y = y + i_n * stride_dy_n + t * stride_dy_t + o_d
-                    b_y = tl.load(p_y, mask=m_t, other=0).to(tl.float32)
-                    b_ys = tl.sigmoid(b_y)
-                    b_dy = b_dy * b_ys * (1 + b_y * (1 - b_ys))
-
-                # Get weight[:, w_idx]
-                b_w_col = tl.load(weight + o_d * W + w_idx, mask=m_d, other=0).to(tl.float32)
-
-                # Accumulate
-                b_dh0 += tl.where(m_t, b_dy * b_w_col, 0)
-
-        # Store dh0[i_n, :, i_w]
-        p_dh0 = dh0 + i_n * D * W + o_d * W + i_w
-        tl.store(p_dh0, b_dh0.to(dh0.dtype.element_ty), mask=m_d)
-
-
-def compute_dh0_triton(
-    dy: torch.Tensor,
-    y: torch.Tensor | None,
-    weight: torch.Tensor,
-    initial_state: torch.Tensor,
-    activation: str | None,
-    cu_seqlens: torch.Tensor | None,
-) -> torch.Tensor:
-    """
-    Compute dh0 (gradient w.r.t. initial_state) using a separate Triton kernel.
-    This is a workaround for Triton compiler bugs on some architectures (e.g., GB200).
-    """
-    D, W = weight.shape
-    N = initial_state.shape[0]
-    T = dy.shape[1]
-
-    # Initialize dh0
-    dh0 = torch.zeros_like(initial_state)
-
-    BD = 32
-    grid = (triton.cdiv(D, BD), N)
-
-    y_to_pass = y if activation in ('swish', 'silu') else None
-    # dy is [B, T, D], stride_n = T*D, stride_t = D
-    stride_dy_n = dy.stride(0)
-    stride_dy_t = dy.stride(1)
-
-    compute_dh0_kernel[grid](
-        dy=dy,
-        y=y_to_pass,
-        weight=weight,
-        dh0=dh0,
-        cu_seqlens=cu_seqlens,
-        stride_dy_n=stride_dy_n,
-        stride_dy_t=stride_dy_t,
-        T=T,
-        D=D,
-        W=W,
-        BD=BD,
-    )
-
-    return dh0
-
-
-def causal_conv1d_bwd(
-    x: torch.Tensor,
-    dy: torch.Tensor,
-    dht: torch.Tensor,
-    weight: torch.Tensor | None = None,
-    bias: torch.Tensor | None = None,
-    residual: torch.Tensor | None = None,
-    initial_state: torch.Tensor | None = None,
-    activation: str | None = None,
-    cu_seqlens: torch.Tensor | None = None,
-    cu_seqlens_cpu: torch.LongTensor | None = None,
-    chunk_indices: torch.LongTensor | None = None,
-):
-    shape = x.shape
-    if x.shape[-1] != weight.shape[0]:
-        x = rearrange(x, 'b t ... -> b t (...)')
-    B, T, D = x.shape
-    W = weight.shape[1] if weight is not None else None
-
-    stride_x_n, stride_x_t, stride_x_d = x.stride()
-
-    BT = min(64, triton.next_power_of_2(triton.cdiv(max(16, B*T), get_multiprocessor_count(x.device.index))))
-    BW = triton.next_power_of_2(W)
-    if chunk_indices is None and (cu_seqlens is not None or cu_seqlens_cpu is not None):
-        chunk_indices = prepare_chunk_indices(cu_seqlens, BT, cu_seqlens_cpu=cu_seqlens_cpu)
-    NT = len(chunk_indices) if cu_seqlens is not None else triton.cdiv(T, BT)
-    NB = triton.cdiv(B*T, 1024)
-
-    y = None
-    if activation is not None:
-        y, _ = causal_conv1d_fwd(
-            x=x,
-            weight=weight,
-            bias=bias,
-            residual=None,
-            initial_state=initial_state,
-            activation=None,
-            cu_seqlens=cu_seqlens,
-            cu_seqlens_cpu=cu_seqlens_cpu,
-            output_final_state=False,
-        )
-    dx = torch.empty_like(x)
-    dw = weight.new_empty(B*NT, *weight.shape, dtype=torch.float) if weight is not None else None
-    db = bias.new_empty(B*NT, *bias.shape, dtype=torch.float) if bias is not None else None
-    dr = dy if residual is not None else None
-
-    stride_dx_n, stride_dx_t, stride_dx_d = dx.stride()
-
-    def grid(meta): return (triton.cdiv(D, meta['BD']), NT, B)
-    causal_conv1d_bwd_kernel[grid](
-        x=x,
-        y=y,
-        weight=weight,
-        initial_state=initial_state,
-        dht=dht,
-        dy=dy,
-        dx=dx,
-        dw=dw,
-        db=db,
-        cu_seqlens=cu_seqlens,
-        chunk_indices=chunk_indices,
-        B=B,
-        T=T,
-        D=D,
-        W=W,
-        BT=BT,
-        BW=BW,
-        NB=NB,
-        stride_x_n=stride_x_n,
-        stride_x_t=stride_x_t,
-        stride_x_d=stride_x_d,
-        stride_dx_n=stride_dx_n,
-        stride_dx_t=stride_dx_t,
-        stride_dx_d=stride_dx_d,
-        ACTIVATION=activation,
-    )
-    if weight is not None:
-        dw = dw.sum(0).to(weight)
-    if bias is not None:
-        db = db.sum(0).to(bias)
-
-    # Compute dh0 using separate Triton kernel to avoid compiler bugs on some architectures (e.g., GB200)
-    dh0 = None
-    if initial_state is not None:
-        dh0 = compute_dh0_triton(
-            dy=dy,
-            y=y,
-            weight=weight,
-            initial_state=initial_state,
-            activation=activation,
-            cu_seqlens=cu_seqlens,
-        )
-
-    return dx.view(shape), dw, db, dr, dh0
-
-
-@triton.heuristics({
-    'USE_INITIAL_STATE': lambda args: args['initial_state'] is not None,
-    'IS_VARLEN': lambda args: args['cu_seqlens'] is not None,
-})
-@triton.jit
-def causal_conv1d_states_fwd_kernel(
-    x,
-    initial_state,
-    final_state,
-    cu_seqlens,
-    T,
-    D,
-    W,
-    stride_x_n,
-    stride_x_t,
-    stride_x_d,
-    BD: tl.constexpr,
-    BW: tl.constexpr,
-    USE_INITIAL_STATE: tl.constexpr,
-    IS_VARLEN: tl.constexpr,
-):
-    i_d, i_n = tl.program_id(0), tl.program_id(1)
-
-    # o_d Shape: [BD]
-    o_d = i_d * BD + tl.arange(0, BD)
-    m_d = o_d < D
-
-    if IS_VARLEN:
-        bos = tl.load(cu_seqlens + i_n).to(tl.int32)
-        eos = tl.load(cu_seqlens + i_n + 1).to(tl.int32)
-        seq_len = eos - bos
-        p_x = x + bos * stride_x_t
-    else:
-        seq_len = T
-        p_x = x + i_n * stride_x_n
-
-    p_x = tl.make_block_ptr(p_x, (seq_len, D), (stride_x_t, stride_x_d), (seq_len - BW, i_d * BD), (BW, BD), (1, 0))
-
-    # b_x Shape: [BW, BD]
-    b_x = tl.load(p_x, boundary_check=(0, 1), padding_option="zero").to(tl.float32)
-
-    if USE_INITIAL_STATE:
-        if seq_len < BW:
-            o_c = W - (BW - seq_len) + tl.arange(0, BW)
-            m_c = (o_c >= 0) & (o_c < W)
-
-            p_init = initial_state + i_n * D*W + o_d[None, :] * W + o_c[:, None]
-            mask_init = m_d[None, :] & m_c[:, None]
-
-            b_cache = tl.load(p_init, mask=mask_init, other=0)
-            b_x += b_cache
-
-    # final_state: [N, D, W] (Channel Major inside sample)
-    # o_w Shape: [BW]
-    o_w = W - BW + tl.arange(0, BW)
-
-    # o_d[:, None] -> [BD, 1]
-    # o_w[None, :] -> [1, BW]
-    # p_final Shape -> [BD, BW]
-    p_final = final_state + i_n * D*W + o_d[:, None] * W + o_w[None, :]
-
-    # m_final Shape -> [BD, BW]
-    m_final = m_d[:, None] & (o_w[None, :] >= 0)
-
-    tl.store(p_final, tl.trans(b_x).to(final_state.dtype.element_ty), mask=m_final)
-
-
-@input_guard(no_guard_contiguous=["x"])
-def causal_conv1d_update_states(
-    x: torch.Tensor,
-    state_len: int,
-    initial_state: torch.Tensor | None = None,
-    cu_seqlens: torch.Tensor | None = None,
-) -> torch.Tensor:
-    if cu_seqlens is not None:
-        N = len(cu_seqlens) - 1
-        if x.dim() == 2:
-            stride_x_n = 0
-            stride_x_t, stride_x_d = x.stride()
-            T = x.shape[0]
-        else:
-            stride_x_n = x.stride(0)
-            stride_x_t, stride_x_d = x.stride(1), x.stride(2)
-            T = x.shape[1]
-        D = x.shape[-1]
-    else:
-        B, T, D = x.shape
-        N = B
-        stride_x_n, stride_x_t, stride_x_d = x.stride()
-
-    W = state_len
-    final_state = torch.empty(N, D, W, dtype=x.dtype, device=x.device)
-
-    BD = min(triton.next_power_of_2(D), 256)
-    BW = triton.next_power_of_2(W)
-
-    grid = (triton.cdiv(D, BD), N)
-
-    causal_conv1d_states_fwd_kernel[grid](
-        x=x,
-        initial_state=initial_state,
-        final_state=final_state,
-        cu_seqlens=cu_seqlens,
-        T=T,
-        D=D,
-        W=W,
-        stride_x_n=stride_x_n,
-        stride_x_t=stride_x_t,
-        stride_x_d=stride_x_d,
-        BW=BW,
-        BD=BD,
-    )
-    return final_state
-
-
-@input_guard(no_guard_contiguous=["x"])
-def causal_conv1d_update(
-    x: torch.Tensor,
-    cache: torch.Tensor,
-    residual: torch.Tensor | None = None,
-    weight: torch.Tensor | None = None,
-    bias: torch.Tensor | None = None,
-    activation: str | None = None,
-) -> torch.Tensor:
-    shape = x.shape
-    if weight is not None and x.shape[-1] != weight.shape[0]:
-        x = rearrange(x, 'b t ... -> b t (...)')
-
-    D = x.shape[-1]
-    N = x.numel() // D
-    W = weight.shape[1] if weight is not None else None
-    BD = 8
-    BW = triton.next_power_of_2(W)
-
-    if x.dim() == 2:
-        # Case: (N, D)
-        stride_x_n = x.stride(0)
-        stride_x_d = x.stride(1)
-    elif x.dim() == 3 and x.shape[0] == 1:
-        # Case: (1, N, D) -> Time=1, Batch=N, Dim=D
-        # Batch 在 dim 1
-        stride_x_n = x.stride(1)
-        stride_x_d = x.stride(2)
-    elif x.dim() == 3:
-        # Case: (N, 1, D) -> Batch=N, Time=1, Dim=D
-        # Batch 在 dim 0
-        stride_x_n = x.stride(0)
-        stride_x_d = x.stride(2)
-    else:
-        # Fallback / Error case
-        raise ValueError(f"Unsupported input shape: {x.shape}")
-
-    y = torch.empty_like(x, memory_format=torch.contiguous_format)
-
-    if y.dim() == 2:
-        stride_y_n, stride_y_d = y.stride(0), y.stride(1)
-    elif y.dim() == 3 and y.shape[0] == 1:
-        stride_y_n, stride_y_d = y.stride(1), y.stride(2)
-    elif y.dim() == 3:
-        stride_y_n, stride_y_d = y.stride(0), y.stride(2)
-
-    def grid(meta): return (triton.cdiv(D, meta['BD']), N)
-
-    causal_conv1d_update_kernel[grid](
-        x=x,
-        cache=cache,
-        residual=residual,
-        y=y,
-        weight=weight,
-        bias=bias,
-        stride_x_n=stride_x_n,
-        stride_x_d=stride_x_d,
-        stride_y_n=stride_y_n,
-        stride_y_d=stride_y_d,
-        D=D,
-        W=W,
-        BD=BD,
-        BW=BW,
-        ACTIVATION=activation,
-        num_warps=STATIC_WARPS,
-    )
-    return y.view(shape), cache
-
-
-class CausalConv1dFunction(torch.autograd.Function):
-
-    @staticmethod
-    @input_guard(no_guard_contiguous=["x"])
-    def forward(
-        ctx,
-        x: torch.Tensor,
-        weight: torch.Tensor | None = None,
-        bias: torch.Tensor | None = None,
-        residual: torch.Tensor | None = None,
-        initial_state: torch.Tensor | None = None,
-        output_final_state: bool | None = False,
-        activation: str | None = None,
-        cu_seqlens: torch.Tensor | None = None,
-        cu_seqlens_cpu: torch.LongTensor | None = None,
-        chunk_indices: torch.LongTensor | None = None,
-    ):
-        ctx.activation = activation
-        ctx.cu_seqlens = cu_seqlens
-        ctx.cu_seqlens_cpu = cu_seqlens_cpu
-        ctx.chunk_indices = chunk_indices
-        ctx.save_for_backward(x, weight, bias, residual, initial_state)
-        y, final_state = causal_conv1d_fwd(
-            x=x,
-            weight=weight,
-            bias=bias,
-            residual=residual,
-            initial_state=initial_state,
-            output_final_state=output_final_state,
-            activation=activation,
-            cu_seqlens=cu_seqlens,
-            cu_seqlens_cpu=cu_seqlens_cpu,
-            chunk_indices=chunk_indices,
-        )
-        return y, final_state
-
-    @staticmethod
-    @input_guard(no_guard_contiguous=["dy"])
-    def backward(ctx, dy: torch.Tensor, dht: torch.Tensor | None = None):
-        x, weight, bias, residual, initial_state = ctx.saved_tensors
-        dx, dw, db, dr, dh0 = causal_conv1d_bwd(
-            x=x,
-            dy=dy,
-            dht=dht,
-            weight=weight,
-            bias=bias,
-            residual=residual,
-            initial_state=initial_state,
-            activation=ctx.activation,
-            cu_seqlens=ctx.cu_seqlens,
-            cu_seqlens_cpu=ctx.cu_seqlens_cpu,
-            chunk_indices=ctx.chunk_indices,
-        )
-        return dx, dw, db, dr, dh0, None, None, None, None, None
-
-
-class FastCausalConv1dFn(torch.autograd.Function):
-    """
-    Mixed-mode (Mix) Causal Convolution Implementation - Combining Triton Forward and CUDA Backward Propagation
-
-    This class implements forward propagation using FLA's Triton kernel, while using the optimized
-    implementation from TriDao's causal_conv1d CUDA package for backward propagation.
-    This hybrid strategy combines the advantages of both technologies:
-
-    - Forward: Uses FLA's Triton implementation, optimized for the FLA framework
-    - Backward: Uses TriDao's causal_conv1d_bwd_function CUDA implementation for faster speed
-
-    Performance Benefits:
-    - CUDA backward implementation is typically faster than the Triton version, reducing training time
-    - Maintains the flexibility and compatibility of forward propagation
-
-    Note:
-    - Input/Output format is (batch, seqlen, dim)
-    - Backward propagation requires causal_conv1d package: pip install causal-conv1d
-    - Supports SILU/Swish activation functions
-    - Current limitations (not yet supported):
-        * output_final_state must be False
-        * initial_states must be None
-        * residual must be None
-    """
-    @staticmethod
-    @input_guard(no_guard_contiguous=["x"])
-    def forward(
-        ctx,
-        x,
-        weight,
-        bias=None,
-        residual: torch.Tensor | None = None,
-        initial_states=None,
-        output_final_state=False,
-        activation=None,
-        cu_seqlens: torch.LongTensor | None = None,
-        cu_seqlens_cpu: torch.LongTensor | None = None,
-        chunk_indices: torch.LongTensor | None = None,
-        seq_idx: torch.LongTensor | None = None,
-    ):
-        if activation not in [None, "silu", "swish"]:
-            raise NotImplementedError("activation must be None, silu, or swish")
-        assert output_final_state is False, "output_final_state must be False for FastCausalConv1dFn"
-        assert initial_states is None, "initial_states must be None for FastCausalConv1dFn"
-        assert residual is None, "residual must be None for FastCausalConv1dFn"
-
-        bias = bias.contiguous() if bias is not None else None
-        if cu_seqlens is not None and seq_idx is None:
-            seq_idx = prepare_sequence_ids(cu_seqlens, cu_seqlens_cpu=cu_seqlens_cpu).to(
-                torch.int32).unsqueeze(0)
-        seq_idx = seq_idx.contiguous() if seq_idx is not None else None
-
-        ctx.activation = activation in ["silu", "swish"]
-        out, _ = causal_conv1d_fwd(
-            x=x,
-            weight=weight,
-            bias=bias,
-            residual=None,
-            initial_state=None,
-            output_final_state=output_final_state,
-            activation=activation,
-            cu_seqlens=cu_seqlens,
-            cu_seqlens_cpu=cu_seqlens_cpu,
-            chunk_indices=chunk_indices,
-        )
-
-        ctx.save_for_backward(x, weight, bias, seq_idx, initial_states)
-        ctx.return_final_states = output_final_state
-        ctx.return_dinitial_states = (
-            initial_states is not None and initial_states.requires_grad
-        )
-        return out, None
-
-    @staticmethod
-    @input_guard
-    def backward(ctx, dout, *args):
-        x, weight, bias, seq_idx, initial_states = ctx.saved_tensors
-        dx = torch.empty_like(x, memory_format=torch.contiguous_format)
-        x = rearrange(x, 'b t d -> b d t')
-        dx = rearrange(dx, 'b t d -> b d t')
-        dout = rearrange(dout, 'b t d -> b d t')
-        dfinal_states = args[0] if ctx.return_final_states else None
-
-        if dout.stride(2) != 1 and dout.stride(1) != 1:
-            dout = dout.contiguous()
-        # The kernel supports passing in a pre-allocated dx (e.g., in case we want to fuse the
-        # backward of conv1d with the backward of chunk).
-        # Here we just pass in None and dx will be allocated in the C++ code.
-        dx, dweight, dbias, dinitial_states = causal_conv1d_bwd_function(
-            x,
-            weight,
-            bias,
-            dout,
-            seq_idx,
-            initial_states,
-            dfinal_states,
-            dx,
-            ctx.return_dinitial_states,
-            ctx.activation,
-        )
-        dx = rearrange(dx, 'b d t -> b t d')
-        return (
-            dx,
-            dweight,
-            dbias if bias is not None else None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-        )
-
-
-def fast_causal_conv1d_fn(
-    x: torch.Tensor,
-    weight: torch.Tensor | None = None,
-    bias: torch.Tensor | None = None,
-    residual: torch.Tensor | None = None,
-    initial_state: torch.Tensor | None = None,
-    output_final_state: bool | None = False,
-    activation: str | None = None,
-    cu_seqlens: torch.Tensor | None = None,
-    cu_seqlens_cpu: torch.LongTensor | None = None,
-    chunk_indices: torch.LongTensor | None = None,
-    seq_idx: torch.LongTensor | None = None,
-):
-    """
-    x: (batch, seqlen, dim)
-    weight: (dim, width)
-    bias: (dim,)
-    seq_idx: (batch, seqlen)
-    initial_states: (batch, dim, width - 1)
-    final_states_out: (batch, dim, width - 1), to be written to
-    activation: either None or "silu" or "swish"
-
-    out: (batch, seqlen, dim)
-    """
-    return FastCausalConv1dFn.apply(
-        x,
-        weight,
-        bias,
-        residual,
-        initial_state,
-        output_final_state,
-        activation,
-        cu_seqlens,
-        cu_seqlens_cpu,
-        chunk_indices,
-        seq_idx,
-    )
-
-
-@input_guard(no_guard_contiguous=["x"])
-def causal_conv1d(
-    x: torch.Tensor,
-    weight: torch.Tensor | None = None,
-    bias: torch.Tensor | None = None,
-    residual: torch.Tensor | None = None,
-    initial_state: torch.Tensor | None = None,
-    output_final_state: bool | None = False,
-    activation: str | None = None,
-    backend: str | None = 'triton',
-    cu_seqlens: torch.Tensor | None = None,
-    cu_seqlens_cpu: torch.LongTensor | None = None,
-    chunk_indices: torch.LongTensor | None = None,
-    cp_context: FLACPContext | None = None,
-    **kwargs,
-):
-    """
-    A causal 1D convolution implementation that powers Mamba/Mamba2 and DeltaNet architectures.
-
-    When a residual connection is provided, this implements the Canon operation
-    described in the paper at https://papers.ssrn.com/sol3/papers.cfm?abstract_id=5240330.
-
-    Args:
-        x (torch.Tensor):
-            Input tensor of shape [B, T, D].
-        weight (Optional[torch.Tensor]):
-            Weight tensor of shape [D, W]. Default: `None`.
-        bias (Optional[torch.Tensor]):
-            Bias tensor of shape [D]. Default: `None`.
-        residual (Optional[torch.Tensor]):
-            Residual tensor of shape [B, T, D]. Default: `None`.
-        initial_state (Optional[torch.Tensor]):
-            Initial state tensor of shape [N, D, W],
-            where `N` is the number of sequences in the batch and `W` is the kernel size.
-            If provided, the initial state is used to initialize the cache. Default: `None`.
-        output_final_state (Optional[bool]):
-            Whether to output the final state of shape [N, D, W]. Default: `False`.
-        activation (Optional[str]):
-            Activations applied to output, only `swish`/`silu` or `None` (i.e., no activation) are supported.
-            Default: `None`.
-        backend (Optional[str]):
-            Specifies the backend to use for the convolution operation. Supported values are `'cuda'` 、 `'triton'` and `'mix'`.
-            Default: `'triton'`.
-        cu_seqlens (Optional[torch.Tensor]):
-            Cumulative sequence lengths (optional)
-        chunk_indices (Optional[torch.LongTensor]):
-            Chunk indices for variable-length sequences (optional)
-
-    Returns:
-        Tuple of (output, final_state).
-        If `output_final_state` is `False`, the final state is `None`.
-    """
-    if cp_context is not None:
-        assert initial_state is None, "Initial state is not supported for CP"
-        assert output_final_state is False, "Output final state is not supported for CP"
-        output = causal_conv1d_cp(
-            x=x,
-            weight=weight,
-            bias=bias,
-            activation=activation,
-            chunk_indices=chunk_indices,
-            cp_context=cp_context,
-        )
-        return output, None
-
-    if backend == 'triton':
-        y, final_state = CausalConv1dFunction.apply(
-            x,
-            weight,
-            bias,
-            residual,
-            initial_state,
-            output_final_state,
-            activation,
-            cu_seqlens,
-            cu_seqlens_cpu,
-            chunk_indices,
-        )
-        return y, final_state
-    elif backend == 'mix':
-        if causal_conv1d_bwd_function is None:
-            raise ImportError(
-                "causal_conv1d is required for backend='mix', but it is not installed. "
-                "Please install it with: pip install causal-conv1d\n"
-                "For more details, see: https://github.com/Dao-AILab/causal-conv1d"
-            )
-        seq_idx = kwargs.get('seq_idx')
-        return fast_causal_conv1d_fn(
-            x,
-            weight,
-            bias,
-            residual,
-            initial_state,
-            output_final_state,
-            activation,
-            cu_seqlens,
-            cu_seqlens_cpu=cu_seqlens_cpu,
-            chunk_indices=chunk_indices,
-            seq_idx=seq_idx,
-        )
-    W = weight.shape[-1]
-    if initial_state is not None:
-        # Case: Has initial_state -> Must be Channel-Last (physically B, T, D)
-        if x.stride(-1) != 1:
-            x = x.contiguous()
-        x = rearrange(x, 'b t d -> b d t')
-    else:
-        # Case: No initial_state -> Prefer Contiguous (physically B, D, T)
-        x = rearrange(x, 'b t d -> b d t').contiguous()
-
-    # check if cu_seqlens and cache are both provided
-    # Sequence index for each token. Used for varlen.
-    # Suppose a batch consists of two sequences with lengths 3 and 4,
-    # seq_idx=[0, 0, 0, 1, 1, 1, 1] for this batch.
-    # NOTE: No need to provide this arg if `cu_seqlens` is passed.
-    # This arg is just for BC, and will be removed in the future.
-    # [B, T]
-    seq_idx = kwargs.get('seq_idx')
-    if cu_seqlens is not None and seq_idx is None:
-        seq_idx = prepare_sequence_ids(cu_seqlens).to(torch.int32).unsqueeze(0)
-
-    # equivalent to:
-    # y = _conv_forward(x, weight, bias)[..., :x.shape[-1]]
-    # if activation is not None:
-    #     y = ACT2FN[activation](x)
-
-    cache, initial_state = initial_state, None
-    if cache is not None:
-        # To make causal-conv1d happy
-        initial_state = (
-            cache[:, :, -(W-1):]   # [N, D, W-1]
-            .transpose(1, 2).contiguous()  # [N, W-1, D] and stride(2)==1
-            .transpose(1, 2)               # [N, D, W-1] and stride(1)==1
-        )
-
-    y = causal_conv1d_fn(
-        x=x,
-        weight=weight,
-        bias=bias,
-        activation=activation,
-        seq_idx=seq_idx,
-        initial_states=initial_state,
-        return_final_states=False,
-    )
-
-    y = rearrange(y, 'b d t -> b t d')
-    if output_final_state:
-        final_state = causal_conv1d_update_states(
-            x=x,
-            state_len=W,
-            initial_state=initial_state,
-            cu_seqlens=cu_seqlens,
-        )
-    if residual is not None:
-        y.add_(residual)
-
-    return y, cache
-
-
-class ShortConvolution(nn.Conv1d):
-    """Short convolution layer for efficient causal convolution operations.
-
-    This class implements a depthwise separable 1D convolution with causal padding,
-    designed for efficient sequence processing. It supports multiple backends (Triton/CUDA)
-    and optional activation functions.
-
-    Args:
-        hidden_size (int): Number of input/output channels (must be equal for depthwise conv)
-        kernel_size (int): Size of the convolution kernel
-        bias (bool, optional): Whether to include learnable bias. Defaults to False.
-        activation (Optional[str], optional): Activation function ('silu' or 'swish'). Defaults to 'silu'.
-        backend (Optional[str], optional): Backend implementation ('triton' or 'cuda'). Defaults to 'triton'.
-        device (Optional[torch.device], optional): Device to place the layer on. Defaults to None.
-        dtype (Optional[torch.dtype], optional): Data type for layer parameters. Defaults to None.
-        **kwargs: Additional keyword arguments (deprecated 'use_fast_conv1d' supported for compatibility)
-
-    Attributes:
-        hidden_size (int): Number of channels
-        activation (Optional[str]): Selected activation function
-        backend (str): Actual backend being used (may differ from input due to availability)
-
-    Note:
-        - Uses depthwise convolution (groups=hidden_size) for efficiency
-        - Applies causal padding (kernel_size-1) to ensure no future information leakage
-        - Falls back to Triton backend if CUDA backend is unavailable
-    """
-
-    def __init__(
-        self,
-        hidden_size: int,
-        kernel_size: int,
-        bias: bool = False,
-        activation: str | None = 'silu',
-        backend: str | None = 'triton',
-        device: torch.device | None = None,
-        dtype: torch.dtype | None = None,
-        **kwargs,
-    ):
-        super().__init__(
-            in_channels=hidden_size,
-            out_channels=hidden_size,
-            kernel_size=kernel_size,
-            groups=hidden_size,
-            bias=bias,
-            padding=kernel_size - 1,
-            device=device,
-            dtype=dtype,
-        )
-
-        self.hidden_size = hidden_size
-        self.activation = None
-
-        if activation is not None:
-            assert activation in ['silu', 'swish'], f"Activation `{activation}` not supported yet."
-            self.activation = activation
-
-        if 'use_fast_conv1d' in kwargs:
-            warnings.warn(
-                "The `use_fast_conv1d` parameter is deprecated and will be ignored. "
-                "Please use the `backend` parameter instead.",
-            )
-        import os
-        self.backend = os.environ.get('FLA_CONV_BACKEND', backend)
-        if backend not in ['cuda', 'triton']:
-            raise ValueError(f"Invalid backend: {backend}, must be one of ['cuda', 'triton']")
-        if backend == 'cuda':
-            if causal_conv1d_fn is None:
-                warnings.warn(
-                    "The `backend` parameter is set to `cuda`, but `causal_conv1d_fn` is not available. "
-                    "Switching to the Triton implementation instead. "
-                    "Consider installing `causal_conv1d` to enable the CUDA backend.",
-                )
-                self.backend = 'triton'
-
-    def extra_repr(self):
-        s = ('{in_channels}, {out_channels}, kernel_size={kernel_size}'
-             ', stride={stride}')
-        if self.padding != (0,) * len(self.padding):
-            s += ', padding={padding}'
-        if self.dilation != (1,) * len(self.dilation):
-            s += ', dilation={dilation}'
-        if self.output_padding != (0,) * len(self.output_padding):
-            s += ', output_padding={output_padding}'
-        if self.groups != 1:
-            s += ', groups={groups}'
-        if self.bias is None:
-            s += ', bias=False'
-        if self.padding_mode != 'zeros':
-            s += ', padding_mode={padding_mode}'
-        if self.activation is not None:
-            s += ', activation={activation}'
-        s += f', backend={self.backend}'
-        return s.format(**self.__dict__)
-
-    def forward(
-        self,
-        x: torch.Tensor,
-        residual: torch.Tensor | None = None,
-        mask: torch.Tensor | None = None,
-        cache: torch.Tensor | None = None,
-        output_final_state: bool = False,
-        cu_seqlens: torch.LongTensor | None = None,
-        chunk_indices: torch.LongTensor | None = None,
-        **kwargs,
-    ) -> tuple[torch.Tensor, torch.Tensor]:
-        """
-        Args:
-            x (`torch.Tensor`):
-                Tensor of shape `[B, T, D]`. `B` must be 1 if `cu_seqlens` is provided.
-            residual (`Optional[torch.Tensor]`):
-                Residual tensor of shape `[B, T, D]`. Default: `None`.
-            mask (`Optional[torch.Tensor]`):
-                Attention mask dealing with padded positions.
-            cache (`Optional[torch.Tensor]`):
-                Previous cache tensor of shape `[N, D, W]`, where `W` is the kernel size.
-                If provided, the cache is updated **inplace**.
-            output_final_state (Optional[bool]):
-                Whether to output the final state of shape `[N, D, W]`. Default: `False`.
-            cu_seqlens (Optional[torch.LongTensor]):
-                Cumulative sequence lengths for each batch. Used for varlen. Default: `None`.
-                Shape: [B+1]
-            chunk_indices (Optional[torch.LongTensor]):
-                Chunk indices for variable-length sequences. Default: `None`.
-
-        Returns:
-            Tensor of shape `[B, T, D]`.
-        """
-
-        B, T, *_ = x.shape
-        N = B if cu_seqlens is None else len(cu_seqlens) - 1
-        if mask is not None:
-            if cu_seqlens is not None:
-                raise ValueError("`mask` and `cu_seqlens` cannot be provided at the same time")
-            x = x.mul_(mask.unsqueeze(-1))
-
-        # in decoding phase, the cache (if provided) is updated inplace
-        if B * T == N:
-            y, cache = self.step(
-                x=x,
-                residual=residual,
-                cache=cache,
-                output_final_state=output_final_state,
-                cu_seqlens=cu_seqlens,
-            )
-            return y, cache
-
-        # cuda backend do not support:
-        # 1. both `cu_seqlens` and `cache` being provided
-        # 2. both `cu_seqlens` and `output_final_state` being provided
-        # and other small issues
-        # to simplify the implementation, we just switch to triton backend
-        if self.backend == 'cuda' and cache is not None:
-            warnings.warn(
-                "The CUDA backend does not support both `cu_seqlens` and `cache` being provided, "
-                "or both `cu_seqlens` and `output_final_state` being provided. "
-                "Switching to the Triton backend instead. ",
-                stacklevel=2,
-            )
-            self.backend = 'triton'
-
-        return causal_conv1d(
-            x=x,
-            weight=rearrange(self.weight, "d 1 w -> d w"),
-            bias=self.bias,
-            residual=residual,
-            initial_state=cache,
-            output_final_state=output_final_state,
-            activation=self.activation,
-            backend=self.backend,
-            cu_seqlens=cu_seqlens,
-            chunk_indices=chunk_indices,
-            **kwargs,
-        )
-
-    def step(
-        self,
-        x: torch.Tensor,
-        residual: torch.Tensor,
-        cache: torch.Tensor,
-        output_final_state: bool = False,
-        cu_seqlens: torch.LongTensor | None = None,
-    ):
-        B, _, D, W = *x.shape, self.kernel_size[0]
-        N = B if cu_seqlens is None else len(cu_seqlens) - 1
-        if output_final_state and cache is None:
-            cache = x.new_zeros(N, D, W)
-        # NOTE: we follow the fast mode that updates the cache in-place
-        if self.backend == 'triton':
-            return causal_conv1d_update(
-                x=x,
-                cache=cache,
-                residual=residual,
-                weight=rearrange(self.weight, "d 1 w -> d w"),
-                bias=self.bias,
-                activation=self.activation,
-            )
-
-        shape = x.shape
-        x = x.squeeze(0) if cu_seqlens is not None else x.squeeze(1)
-        # equivalent to:
-        # cache.copy_(cache.roll(shifts=-1, dims=-1))
-        # cache[:, :, -1] = x
-        # y = torch.sum(cache * rearrange(self.weight, "d 1 w -> d w"), dim=-1)
-        y = causal_conv1d_update_cuda(
-            x=x,
-            conv_state=cache,
-            weight=rearrange(self.weight, "d 1 w -> d w"),
-            bias=self.bias,
-            activation=self.activation,
-        )
-        y = y.view(shape)
-        if residual is not None:
-            y.add_(residual)
-        return y, cache
-
-    @property
-    def state_size(self) -> int:
-        return self.hidden_size * self.kernel_size
-
-
-def fft_conv(u, k, dropout_mask, gelu=True, k_rev=None):
-    seqlen = u.shape[-1]
-    fft_size = 2 * seqlen
-    k_f = torch.fft.rfft(k, n=fft_size) / fft_size
-    if k_rev is not None:
-        k_rev_f = torch.fft.rfft(k_rev, n=fft_size) / fft_size
-        k_f = k_f + k_rev_f.conj()
-    u_f = torch.fft.rfft(u.to(dtype=k.dtype), n=fft_size)
-
-    if len(u.shape) > 3:
-        k_f = k_f.unsqueeze(1)
-    y = torch.fft.irfft(u_f * k_f, n=fft_size, norm="forward")[..., :seqlen]
-
-    out = y + u
-    if gelu:
-        out = F.gelu(out)
-    if dropout_mask is not None:
-        return (out * rearrange(dropout_mask, "b H -> b H 1")).to(dtype=u.dtype)
-    else:
-        return out.to(dtype=u.dtype)
-
-
-class LongConvolution(nn.Module):
-    """
-    LongConvolution applies a convolution operation on the input tensor using a fixed
-    filter of length max_len.
-    The filter is learned during training and is applied using FFT convolution.
-
-    Args:
-        hidden_size (int): The number of expected features in the input and output.
-        max_len (int): The maximum sequence length.
-
-    Returns:
-        y: [batch_size, seq_len, hidden_size] tensor
-    """
-
-    def __init__(
-        self,
-        hidden_size: int,
-        max_len: int,
-        **kwargs,
-    ):
-        """
-        Initializes the LongConvolution module.
-        Args:
-            hidden_size (int): The number of expected features in the input and output.
-            max_len (int): The maximum sequence length.
-        """
-        super().__init__()
-        self.hidden_size = hidden_size
-        self.filter = nn.Parameter(torch.randn(self.hidden_size, max_len), requires_grad=True)
-
-    def forward(self, x: torch.Tensor, *args, **kwargs):
-        """
-        Applies the LongConvolution operation on the input tensor.
-        Args:
-            x: [batch_size, seq_len, hidden_size] tensor
-        Returns:
-            y: [batch_size, seq_len, hidden_size] tensor
-        """
-        x = x.transpose(1, 2)
-        y = fft_conv(x, self.filter, dropout_mask=None, gelu=False)
-        y = y.transpose(1, 2)
-        return y.to(dtype=x.dtype)
-
-
-class PositionalEmbedding(nn.Module):
-    def __init__(self, emb_dim: int, seq_len: int, **kwargs):
-        """Complex exponential positional embeddings for implicit long convolution filters."""
-        super().__init__()
-
-        self.seq_len = seq_len
-        # The time embedding fed to the filteres is normalized so that t_f = 1
-        t = torch.linspace(0, 1, self.seq_len)[None, :, None]  # 1, L, 1
-
-        if emb_dim > 1:
-            bands = (emb_dim - 1) // 2
-        # To compute the right embeddings we use the "proper" linspace
-        t_rescaled = torch.linspace(0, seq_len - 1, seq_len)[None, :, None]
-        w = 2 * math.pi * t_rescaled / seq_len  # 1, L, 1
-
-        f = torch.linspace(1e-4, bands - 1, bands)[None, None]
-        z = torch.exp(-1j * f * w)
-        z = torch.cat([t, z.real, z.imag], dim=-1)
-        self.z = nn.Parameter(z, requires_grad=False)
-
-    def forward(self, L):
-        return self.z[:, :L]
-
-
-class ImplicitLongConvolution(nn.Module):
-    """
-    Long convolution with implicit filter parameterized by an MLP.
-
-    Args:
-        hidden_size (int):
-            The number of expected features in the input and output.
-        max_len (int):
-            The maximum sequence length.
-        d_emb (Optional[int]):
-            The dimension of the positional embeddings. Must be odd and greater or equal to 3 (time, sine and cosine).
-            Defaults to 3.
-        d_hidden (Optional[int]):
-            The number of features in the hidden layer of the MLP. Defaults to 16.
-
-    Attributes:
-        pos_emb (`PositionalEmbedding`): The positional embedding layer.
-        mlp (`nn.Sequential`): The MLP that parameterizes the implicit filter.
-
-    """
-
-    def __init__(
-        self,
-        hidden_size: int,
-        max_len: int,
-        d_emb: int = 3,
-        d_hidden: int = 16,
-        **kwargs,
-    ):
-        """
-        Long convolution with implicit filter parameterized by an MLP.
-
-
-        """
-        super().__init__()
-        self.hidden_size = hidden_size
-        self.d_emb = d_emb
-
-        assert (
-            d_emb % 2 != 0 and d_emb >= 3
-        ), "d_emb must be odd and greater or equal to 3 (time, sine and cosine)"
-        self.pos_emb = PositionalEmbedding(d_emb, max_len)
-
-        # final linear layer
-        self.mlp = nn.Sequential(
-            nn.Linear(d_emb, d_hidden),
-            torch.nn.ReLU(),
-            nn.Linear(d_hidden, hidden_size),
-        )
-
-    def filter(self, seq_len: int, *args, **kwargs):
-        return self.mlp(self.pos_emb(seq_len)).transpose(1, 2)
-
-    def forward(self, x: torch.Tensor, *args, **kwargs):
-        """
-        Args:
-            x: [batch_size, seq_len, hidden_size] tensor
-
-        Returns:
-            y: [batch_size, seq_len, hidden_size] tensor
-        """
-        x = x.transpose(1, 2)
-        k = self.filter(x.shape[-1])
-        y = fft_conv(x, k, dropout_mask=None, gelu=False)
-
-        y = y.transpose(1, 2)
-        return y.to(dtype=x.dtype)
-
-
-# CP Related Conv1d Functions
-
-class CausalConv1dFunctionCP(torch.autograd.Function):
-    """
-    Context Parallel version of CausalConv1dFunction.
-
-    Forward:
-        1. Get tails from previous rank to construct initial_state
-        2. Call causal_conv1d_fwd
-
-    Backward:
-        1. Call causal_conv1d_bwd to get dx
-        2. Sync communication: add next rank's first W-1 token gradients to current rank's last W-1 tokens
-    """
-
-    @staticmethod
-    def _prepare_initial_state_for_cp(
-        x: torch.Tensor,
-        weight: torch.Tensor,
-        cu_seqlens: torch.Tensor | None,
-        context: FLACPContext,
-        group: dist.ProcessGroup | None,
-    ) -> torch.Tensor | None:
-        """Prepare initial_state for CP forward pass by communicating with previous rank.
-
-        Args:
-            x: Input tensor of shape [1, T, D]
-            weight: Weight tensor of shape [D, W]
-            cu_seqlens: Cumulative sequence lengths
-            context: CP context
-            group: Process group for communication
-
-        Returns:
-            initial_state: Initial state tensor of shape [N, D, W] or None
-        """
-        if group is None:
-            return None
-
-        W = weight.shape[-1]  # weight: [D, W]
-        D = weight.shape[0]
-        initial_state = None
-        if not context.is_first_rank:
-            # Non-first rank needs initial_state
-            assert x.dim() == 3 and x.shape[0] == 1, f"CP requires [1, T, D], got {x.shape}"
-            x_2d = x.squeeze(0)  # [T, D]
-            tails = x_2d[-(W-1):].contiguous()  # [W-1, D]
-            heads = conv_cp_send_recv_fwd(tails, group)  # [W-1, D]
-            # Construct initial_state: [N, D, W]
-            N = len(cu_seqlens) - 1
-            initial_state = torch.zeros(N, D, W, device=x.device, dtype=x.dtype)
-            valid_len = min(W - 1, context.pre_num_conv_tokens)
-            if valid_len > 0:
-                # heads[-valid_len:]: [valid_len, D] -> [D, valid_len]
-                initial_state[0, :, -valid_len:] = heads[-valid_len:].T
-        else:
-            # First rank also needs to participate in communication (send tails)
-            x_2d = x.squeeze(0)
-            tails = x_2d[-(W-1):].contiguous()
-            _ = conv_cp_send_recv_fwd(tails, group)  # Send but don't use
-
-        return initial_state
-
-    @staticmethod
-    def _correct_dx_for_cp(
-        dx: torch.Tensor,
-        dh0: torch.Tensor | None,
-        W: int,
-        group: dist.ProcessGroup | None,
-        is_first_rank: bool,
-    ) -> None:
-        """Correct dx gradients for CP backward pass by communicating with next rank.
-
-        Args:
-            dx: Gradient tensor to be corrected, shape [1, T, D]
-            dh0: Gradient w.r.t. initial_state, shape [N, D, W] or None
-            W: Kernel size
-            group: Process group for communication
-            is_first_rank: Whether this is the first rank in the sequence's processing chain
-        """
-        if group is None:
-            return
-
-        D = dx.shape[-1]
-        # dh0: [N, D, W] or None
-        # We only care about the first sequence's initial_state gradient
-        if dh0 is not None:
-            # Get first sequence's d_initial_state: [D, W] -> last W-1 cols -> [D, W-1] -> [W-1, D]
-            d_initial_state = dh0[0, :, -(W-1):].T.contiguous()  # [W-1, D]
-        else:
-            # dh0 is None only when this is the first rank (no initial_state needed)
-            assert is_first_rank, "dh0 should not be None when is_first_rank=False"
-            d_initial_state = torch.zeros(W-1, D, device=dx.device, dtype=dx.dtype)
-        # Sync communication: send d_initial_state to previous rank, receive from next rank
-        recv_d_init = conv_cp_send_recv_bwd(d_initial_state, group)  # [W-1, D]
-        # Add to current rank's last W-1 tokens (these tokens are used as initial_state by next rank)
-        dx[0, -(W-1):, :].add_(recv_d_init)
-
-    @staticmethod
-    def forward(
-        ctx,
-        x: torch.Tensor,
-        weight: torch.Tensor,
-        bias: torch.Tensor | None,
-        activation: str | None,
-        chunk_indices: torch.Tensor | None,
-        cp_context: FLACPContext | None,
-    ):
-        if cp_context is None:
-            raise ValueError("cp_context must be provided for CausalConv1dFunctionCP")
-        cu_seqlens = cp_context.cu_seqlens
-        cu_seqlens_cpu = cp_context.cu_seqlens_cpu
-        group = cp_context.group
-
-        # Get kernel_size
-        W = weight.shape[-1]  # weight: [D, W]
-        # Prepare initial_state for CP
-        initial_state = CausalConv1dFunctionCP._prepare_initial_state_for_cp(
-            x=x,
-            weight=weight,
-            cu_seqlens=cu_seqlens,
-            context=cp_context,
-            group=group,
-        )
-
-        ctx.save_for_backward(x, weight, bias, initial_state)
-        ctx.activation = activation
-        ctx.cu_seqlens = cu_seqlens
-        ctx.cu_seqlens_cpu = cu_seqlens_cpu
-        ctx.chunk_indices = chunk_indices
-        ctx.group = group
-        ctx.W = W
-        ctx.is_first_rank = cp_context.is_first_rank
-
-        # Call original forward
-        y, _ = causal_conv1d_fwd(
-            x=x,
-            weight=weight,
-            bias=bias,
-            residual=None,
-            initial_state=initial_state,
-            output_final_state=False,
-            activation=activation,
-            cu_seqlens=cu_seqlens,
-            cu_seqlens_cpu=cu_seqlens_cpu,
-            chunk_indices=chunk_indices,
-        )
-
-        return y
-
-    @staticmethod
-    def backward(ctx, dy: torch.Tensor):
-        x, weight, bias, initial_state = ctx.saved_tensors
-        group = ctx.group
-        W = ctx.W
-
-        # Call original backward
-        dx, dw, db, _, dh0 = causal_conv1d_bwd(
-            x=x,
-            dy=dy,
-            dht=None,
-            weight=weight,
-            bias=bias,
-            residual=None,
-            initial_state=initial_state,
-            activation=ctx.activation,
-            cu_seqlens=ctx.cu_seqlens,
-            cu_seqlens_cpu=ctx.cu_seqlens_cpu,
-            chunk_indices=ctx.chunk_indices,
-        )
-
-        # Correct dx gradients for CP
-        CausalConv1dFunctionCP._correct_dx_for_cp(
-            dx=dx,
-            dh0=dh0,
-            W=W,
-            group=group,
-            is_first_rank=ctx.is_first_rank,
-        )
-
-        return dx, dw, db, None, None, None
-
-
-def causal_conv1d_cp(
-    x: torch.Tensor,
-    weight: torch.Tensor,
-    bias: torch.Tensor | None = None,
-    activation: str | None = None,
-    chunk_indices: torch.Tensor | None = None,
-    cp_context: FLACPContext | None = None,
-):
-    """
-    Context Parallel version of causal_conv1d.
-
-    Automatically handles communication in CP environment:
-    - Forward: get initial_state from previous rank
-    - Backward: correct dx gradients
-
-    Args:
-        x: Input tensor of shape [1, T, D]
-        weight: Weight tensor of shape [D, W]
-        bias: Bias tensor of shape [D] or None
-        activation: Activation function name or None
-        chunk_indices: Chunk indices for variable-length sequences
-        cp_context: CP context (required for CP mode)
-    """
-    if cp_context is None:
-        raise ValueError("cp_context must be provided for causal_conv1d_cp")
-
-    assert cp_context.conv1d_kernel_size is not None, "conv1d_kernel_size must be provided for causal_conv1d_cp"
-    assert cp_context.cu_seqlens is not None, "cu_seqlens must be provided for causal_conv1d_cp"
-    return CausalConv1dFunctionCP.apply(
-        x, weight, bias, activation,
-        chunk_indices, cp_context
-    )
+__all__ = [
+    'CausalConv1dFunction',
+    'CausalConv1dFunctionCP',
+    'FastCausalConv1dFn',
+    'ImplicitLongConvolution',
+    'LongConvolution',
+    'PositionalEmbedding',
+    'ShortConvolution',
+    'causal_conv1d',
+    'causal_conv1d_bwd',
+    'causal_conv1d_cp',
+    'causal_conv1d_fwd',
+    'causal_conv1d_update',
+    'causal_conv1d_update_states',
+    'fast_causal_conv1d_fn',
+    'fft_conv',
+]
